# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Assesify API
  version: "1.0.0"
  description: >
    Hand-authored OpenAPI schema describing the current Assesify Flask API
    surface (spec.md §4). Used as the source for `openapi-typescript` to
    generate frontend TS types (spec.md §7.3). Reflects the API as it exists
    today, not an aspirational redesign — kept in sync manually since the
    backend does not (yet) generate this from code.
servers:
  - url: /api/v1
    description: Versioned API root (also mounted at /api for legacy routes)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    User:
      type: object
      required:
        - id
        - email
        - full_name
        - is_teacher
        - is_admin
        - streak
        - diamonds
        - health
        - created_at
      properties:
        id:
          type: integer
        email:
          type: string
        full_name:
          type: string
        is_teacher:
          type: boolean
        is_admin:
          type: boolean
        major:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        profile_pic:
          type: string
          nullable: true
        karmayogi_user_id:
          type: string
          nullable: true
        streak:
          type: integer
        diamonds:
          type: integer
        health:
          type: integer
        last_active_date:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
    LoginResponse:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
    RegisterRequest:
      type: object
      required: [email, full_name, password]
      properties:
        email:
          type: string
        full_name:
          type: string
        password:
          type: string
        is_teacher:
          type: boolean
          default: false
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        full_name:
          type: string
        major:
          type: string
        location:
          type: string
    KarmayogiLinkRequest:
      type: object
      required: [karmayogi_user_id]
      properties:
        karmayogi_user_id:
          type: string
    Material:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        mime_type:
          type: string
        created_at:
          type: string
          format: date-time
        extracted_text:
          type: string
...
    GenerateMaterialQuizRequest:
      type: object
      properties:
        num_questions:
          type: integer
          default: 10
        difficulty:
          type: string
          default: mixed
        competency_tags:
          type: array
          items:
            type: string
    GenerateMaterialQuizResponse:
      type: object
      properties:
        id:
          type: integer
        material_id:
          type: integer
        title:
          type: string
        num_questions:
          type: integer
        competency_tags:
          type: array
          items:
            type: string
...
paths:
  /auth/register:
    post:
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
        "400":
          description: Validation error / email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      operationId: refreshToken
      security:
        - bearerAuth: []
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
  /auth/profile:
    get:
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/update-profile:
    put:
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
  /auth/karmayogi/link:
    post:
      operationId: linkKarmayogi
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KarmayogiLinkRequest"
      responses:
        "200":
          description: Linked
          content:
            application/json:
              schema:
                type: object
                properties:
                  karmayogi_user_id:
                    type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/karmayogi/authorize:
    post:
      operationId: startKarmayogiAuthorization
      summary: Begin the OAuth2 authorization-code + PKCE identity-link flow.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                redirect_uri:
                  type: string
                  description: Overrides KARMAYOGI_REDIRECT_URI for this request.
      responses:
        "200":
          description: Authorization URL and anti-CSRF state
          content:
            application/json:
              schema:
                type: object
                required: [authorization_url, state, code_challenge_method, expires_at]
                properties:
                  authorization_url:
                    type: string
                  state:
                    type: string
                  code_challenge_method:
                    type: string
                    enum: [S256]
                  expires_at:
                    type: string
        "503":
          description: Karmayogi OAuth is not configured on this deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/karmayogi/callback:
    post:
      operationId: completeKarmayogiAuthorization
      summary: Exchange the authorization code and link the Karmayogi identity.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, state]
              properties:
                code:
                  type: string
                state:
                  type: string
      responses:
        "200":
          description: Identity linked
          content:
            application/json:
              schema:
                type: object
                required: [karmayogi_user_id, linked_at]
                properties:
                  karmayogi_user_id:
                    type: string
                  linked_at:
                    type: string
                  scope:
                    type: string
                    nullable: true
                  token_type:
                    type: string
                    nullable: true
        "400":
          description: Invalid, expired, or already-used state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Karmayogi rejected the exchange or returned no identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Karmayogi OAuth is not configured on this deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /materials:
    post:
      operationId: createMaterial
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
      responses:
        "201":
          description: Material created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "403":
          description: Forbidden
        "422":
          description: Extraction error
  /materials/{material_id}:
    get:
      operationId: getMaterial
      security:
        - bearerAuth: []
      parameters:
        - name: material_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material with text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{material_id}/generate-quiz:
    post:
      operationId: generateMaterialQuiz
      security:
        - bearerAuth: []
      parameters:
        - name: material_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMaterialQuizRequest"
      responses:
        "201":
          description: Quiz generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateMaterialQuizResponse"
...
  /teacher/materials:
    post:
      operationId: uploadTeacherMaterial
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                subject:
                  type: string
                numQuestions:
                  type: integer
                difficulty:
                  type: string
      responses:
        "201":
          description: Quiz generated and saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  quiz_id:
                    type: integer
                  lesson_id:
                    type: integer
                  title:
                    type: string
                  num_questions:
                    type: integer
...